Add list of research search mission queries
This list is a handy way of encapsulating the thought process of potential users of this tool. This is based on the idea that, I had a need, these searches reflect my thoughts as I look for a solution. Once the tool is built, we need to cater to people who are following in my footsteps and therefore be able to provide answers to people who are performing similar searches.